second commit

+++ calculator.java
@@ -4,7 +4,7 @@ class Main {
   public static void main(String[] args) {
 
     char operator;
-    double  number1, number2, result;
+    int  number1, number2, result;
 
     // create an object of Scanner class
     Scanner input = new Scanner(System.in);
